Shim the remove and replaceWith functions to know when the placeholder element is removed.
This is a little bit risky, but the idea is that we know that the placeholder element is either removed or replacedWith due to the definition of prepare. We can then shim these functions to avoid having a separate function call in the prepare functionality.

The benefit here is that we can avoid having to call the separate function for all elements that isn't placeholders, which should be a subtantial amount, which should balance out the risk of introducing bugs around this behavior.